split out delegator reward tests into own file
